// src/app/pages/hotel/amenities-hotel/extra-bed/shared/extra-bed-enum.ts
export enum ExtraBedEnum {
    titleComponent = 'Extra Bed Options',
    descriptiveText = 'These are the options for beds that can be added upon request',
    
    guest_extraBeds_cot = 'Children up to 2 years old in cribs',
    guest_extraBeds_children = 'Children', 
    guest_extraBeds_adult = 'Adults',
    guestExtraBedsCotId = 1,
    guestExtraBedsChildrenId = 2, 
    guestExtraBedsAdultId = 3,

    hasExtraBed = 'hasExtraBed',
    extraBedQuantity = 'extraBedQuantity',
    guestType = 'guestType',
    cotPrice = 'cotPrice',
    extraBedChildrenAge = 'extraBedChildrenAge',
    extraBedChildrenPrice = 'extraBedChildrenPrice',
    extraBedAdultPrice = 'extraBedAdultPrice',

    defaultDropdownSelection = 1,
    trueString = 'true',
    falseString = 'false',
    priceInputMaxSize = 5,
    emptyString = '',
    model = 'model',

    regexNumbers='[1-9][0-9]*$',

    getItemsErrorMessage = 'There was an error trying to get the Age of Children',
}